refactor networks service (#28)
