+++ networks/service/friend/friend.go
@@ -1,0 +1,1 @@
+package friend
