+++ networks/service/service.go
@@ -1,10 +1,8 @@
 package service
 
-import "github.com/dwarvesf/yggdrasil/follow/service/follow"
+import "github.com/dwarvesf/yggdrasil/networks/service/follow"
 
 // Service ...
 type Service struct {
 	FollowService follow.Service
 }
-
-
