+++ notification/.env
@@ -1,2 +1,1 @@
-CREDENTIAL_FILE="yggdrasil-216009-firebase-adminsdk-cdimi-a653a2d152.json"
 PORT=65535
